requirements.txt
flask==2.0.1
cryptography==3.4.7
pycryptodome==3.10.1
requests==2.26.0
python-dotenv==0.19.0
pynput==1.7.3
opencv-python==4.5.3.56
numpy==1.21.2
pillow==8.3.2
psutil==5.8.0
pywin32==303
pyautogui==0.9.53
keyboard==0.13.5
pycryptodomex==3.10.1
dnspython==2.1.0
pywin32-ctypes==0.2.2
pywin32-security==0.1.0
pywin32-netcon==0.1.0
pywin32-service==0.1.0
pywin32-serviceutil==0.1.0 